// Sources/AuroraStruct3D/Aurora.Vben/apps/web-ele/src/api-client/types.gen.ts
export type ApplicationAuthConfigurationDto = {
    grantedPolicies?: {
        [key: string]: (boolean);
    } | null;
};

export type ApplicationConfigurationDto = {
    localization?: ApplicationLocalizationConfigurationDto;
    auth?: ApplicationAuthConfigurationDto;
    setting?: ApplicationSettingConfigurationDto;
    currentUser?: CurrentUserDto;
    features?: ApplicationFeatureConfigurationDto;
    globalFeatures?: ApplicationGlobalFeatureConfigurationDto;
    multiTenancy?: MultiTenancyInfoDto;
    currentTenant?: CurrentTenantDto;
    timing?: TimingDto;
    clock?: ClockDto;
    objectExtensions?: ObjectExtensionsDto;
    extraProperties?: {
        [key: string]: unknown;
    } | null;
};

export type ApplicationFeatureConfigurationDto = {
    values?: {
        [key: string]: ((string) | null);
    } | null;
};

export type ApplicationGlobalFeatureConfigurationDto = {
    enabledFeatures?: Array<(string)> | null;
};

export type ApplicationLocalizationConfigurationDto = {
    values?: {
        [key: string]: {
            [key: string]: (string);
        };
    } | null;
    resources?: {
        [key: string]: ApplicationLocalizationResourceDto;
    } | null;
    languages?: Array<LanguageInfo> | null;
    currentCulture?: CurrentCultureDto;
    defaultResourceName?: (string) | null;
    languagesMap?: {
        [key: string]: Array<NameValue>;
    } | null;
    languageFilesMap?: {
        [key: string]: Array<NameValue>;
    } | null;
};

export type ApplicationLocalizationDto = {
    resources?: {
        [key: string]: ApplicationLocalizationResourceDto;
    } | null;
    currentCulture?: CurrentCultureDto;
};

export type ApplicationLocalizationResourceDto = {
    texts?: {
        [key: string]: (string);
    } | null;
    baseResources?: Array<(string)> | null;
};

export type ApplicationOidcConfigurationDto = {
    /**
     * 是否启用
     */
    enabled?: boolean;
    oidcConfiguration?: Array<OidcConfiguration> | null;
};

export type ApplicationSettingConfigurationDto = {
    values?: {
        [key: string]: ((string) | null);
    } | null;
};

export type ChangePasswordInput = {
    currentPassword?: (string) | null;
    newPassword: string;
};

export type ClockDto = {
    kind?: (string) | null;
};
// ...
export type CurrentCultureDto = {
    displayName?: (string) | null;
    englishName?: (string) | null;
    threeLetterIsoLanguageName?: (string) | null;
    twoLetterIsoLanguageName?: (string) | null;
    isRightToLeft?: boolean;
    cultureName?: (string) | null;
    name?: (string) | null;
    nativeName?: (string) | null;
    dateTimeFormat?: DateTimeFormatDto;
};

export type CurrentTenantDto = {
    id?: (string) | null;
    name?: (string) | null;
    isAvailable?: boolean;
};

export type CurrentUserDto = {
    isAuthenticated?: boolean;
    id?: (string) | null;
    tenantId?: (string) | null;
    impersonatorUserId?: (string) | null;
    impersonatorTenantId?: (string) | null;
    impersonatorUserName?: (string) | null;
    impersonatorTenantName?: (string) | null;
    userName?: (string) | null;
    name?: (string) | null;
    surName?: (string) | null;
    email?: (string) | null;
    emailVerified?: boolean;
    phoneNumber?: (string) | null;
    phoneNumberVerified?: boolean;
    roles?: Array<(string)> | null;
    sessionId?: (string) | null;
};
// ...
export type ExtensionEnumDto = {
    fields?: Array<ExtensionEnumFieldDto> | null;
    localizationResource?: (string) | null;
};
// ...
export type LanguageInfo = {
    cultureName?: (string) | null;
    uiCultureName?: (string) | null;
    displayName?: (string) | null;
    readonly twoLetterISOLanguageName?: (string) | null;
};
// ...
export type ModuleExtensionDto = {
    entities?: {
        [key: string]: EntityExtensionDto;
    } | null;
    configuration?: {
        [key: string]: unknown;
    } | null;
};

export type MultiTenancyInfoDto = {
    isEnabled?: boolean;
};

export type MyProfileOutput = {
    tenantId?: (string) | null;
    userName?: (string) | null;
    name?: (string) | null;
    surname?: (string) | null;
    email?: (string) | null;
    phoneNumber?: (string) | null;
    isActive?: boolean;
    twoFactorEnabled?: boolean;
};

export type NameValue = {
    name?: (string) | null;
    value?: (string) | null;
};

export type NeedChangePasswordOutput = {
    needChangePassword?: boolean;
    message?: (string) | null;
};

export type ObjectExtensionsDto = {
    modules?: {
        [key: string]: ModuleExtensionDto;
    } | null;
    enums?: {
        [key: string]: ExtensionEnumDto;
    } | null;
};
// ...
export type TimeZone = {
    iana?: IanaTimeZone;
    windows?: WindowsTimeZone;
};

export type TimingDto = {
    timeZone?: TimeZone;
};